"Refactor message packing to UsbDevicePrimitive
Moved the `pack_short_message` and `pack_long_message` utility functions from `ThorlabsDevice` to `UsbDevicePrimitive`. Streamlined hardware info retrieval by consolidating logic in `UsbDevicePrimitive`."
