+++ src/devices/usb_device_primitive.rs
@@ -11,14 +11,19 @@ communicating with the underlying USB device.
 Notes:
 */
 
-use crate::env::{BUFFER_SIZE, IN_ENDPOINT, OUT_ENDPOINT, POLL_READ_INTERVAL, SHORT_TIMEOUT};
-use crate::messages::{get_length, get_waiting_sender, MsgFormat};
-use crate::Error;
+use crate::env::{
+    BUFFER_SIZE, DEST, IN_ENDPOINT, LONG_TIMEOUT, OUT_ENDPOINT, POLL_READ_INTERVAL, SHORT_TIMEOUT,
+    SOURCE,
+};
+use crate::error::{DeviceError, Error};
+use crate::messages::ChannelStatus::{New, Sub};
+use crate::messages::{get_length, get_rx_new_or_sub, get_waiting_sender, MsgFormat};
 use rusb::{DeviceDescriptor, DeviceHandle, GlobalContext, Language};
 use std::collections::VecDeque;
 use std::sync::Arc;
 use std::time::Duration;
 use tokio::sync::oneshot::{channel, error::TryRecvError, Receiver, Sender};
+use tokio::time::timeout;
 
 /// # UsbDevicePrimitive
 /// This struct provides a wrapper around the rusb `DeviceHandle` struct,
@@ -29,13 +34,13 @@ use tokio::sync::oneshot::{channel, error::TryRecvError, Receiver, Sender};
 /// # Example
 /// ```rust
 /// use thormotion::devices::UsbDevicePrimitive;
-/// use thormotion::enumerate::get_device;
+/// use thormotion::enumerate::get_device_primitive;
 /// use thormotion::Error;
 ///
 /// fn main() -> Result<(), Error> {
 ///     // Initialize USB device
 ///     let serial_number: &str = "USB123456";
-///     let device: UsbDevicePrimitive = get_device(serial_number)?;
+///     let device: UsbDevicePrimitive = get_device_primitive(serial_number)?;
 ///     
 ///     // The device is now initialised and ready for communication
 ///     Ok(())
@@ -125,7 +130,7 @@ impl UsbDevicePrimitive {
     /// an `Error::DeviceError` containing the device's serial number is returned.
     pub(crate) fn port_write(&self, data: MsgFormat) -> Result<(), Error> {
         if data.len() != self.handle.write_bulk(OUT_ENDPOINT, &data, SHORT_TIMEOUT)? {
-            return Err(Error::DeviceError(format!(
+            return Err(DeviceError(format!(
                 // todo ?make `serial_number` struct which implements `From<T>` so it can be made from numbers and strings, also implements `Display` so it automatically prints "(serial number: #######)"?
                 "Failed to write correct number of bytes to device (serial number: {})",
                 self.serial_number,
@@ -238,5 +243,81 @@ impl UsbDevicePrimitive {
             Ok::<(), Error>(())
         });
         Ok(())
+    }
+
+    /// # Pack Functions
+    ///
+    /// The Thorlabs APT communication protocol uses a fixed length 6-byte message header, which
+    /// may be followed by a variable-length data packet.
+    /// For simple commands, the 6-byte message header is sufficient to convey the entire command.
+    /// For more complex commands (e.g. commands where a set of parameters needs to be passed
+    /// to the device) the 6-byte header is insufficient and must be followed by a data packet.
+    ///
+    /// The `MsgFormat` enum is used to wrap the bytes of a message and indicate whether the
+    /// message is `Short` (six byte header only) or `Long` (six byte header plus variable length
+    /// data package).
+    ///
+    /// The `pack_short_message()` and `pack_long_message()` helper functions are implemented to
+    /// simplify message formatting and enforce consistency with the APT protocol.
+    pub(crate) fn pack_short_message(id: [u8; 2], param1: u8, param2: u8) -> MsgFormat {
+        MsgFormat::Short([id[0], id[1], param1, param2, DEST, SOURCE])
+    }
+
+    pub(crate) fn pack_long_message(id: [u8; 2], length: usize) -> MsgFormat {
+        let mut data: Vec<u8> = Vec::with_capacity(length);
+        data.extend(id);
+        data.extend(((length - 6) as u16).to_le_bytes());
+        data.push(DEST | 0x80);
+        data.push(SOURCE);
+        MsgFormat::Long(data)
+    }
+
+    /// # HW_REQ_INFO (0x0005)
+    ///
+    /// **Function implemented from Thorlabs APT protocol**
+    ///
+    /// This function is used to request hardware information from the controller.
+    ///
+    /// Message ID: 0x0005
+    ///
+    /// Message length: 6 bytes (header only)
+    ///
+    /// # Response
+    ///
+    /// The controller will send a `HW_GET_INFO (0x0006)` message in response, which
+    /// is then parsed into the component values and packaged into a tuple.
+    ///
+    /// Response ID: 0x0006
+    ///
+    /// Response length: 90 bytes (6-byte header followed by 84-byte data packet)
+    async fn hw_req_info(&self) -> Result<(u32, String, u16, String, u16, u16, u16), Error> {
+        const ID: [u8; 2] = [0x00, 0x05];
+        let mut rx = match get_rx_new_or_sub(ID)? {
+            Sub(rx) => rx,
+            New(rx) => {
+                let data = Self::pack_short_message(ID, 0, 0);
+                self.port_write(data)?;
+                rx
+            }
+        };
+        let response = timeout(LONG_TIMEOUT, rx.recv()).await??;
+        let serial_number = u32::from_le_bytes(response[6..10].try_into()?);
+        let model_number = String::from_utf8_lossy(&response[10..18]).to_string();
+        let hardware_type = u16::from_le_bytes(response[18..20].try_into()?);
+        let firmware_minor = u8::from_le_bytes(response[20..21].try_into()?);
+        let firmware_interim = u8::from_le_bytes(response[21..22].try_into()?);
+        let firmware_major = u8::from_le_bytes(response[22..23].try_into()?);
+        let hardware_version = u16::from_le_bytes(response[84..86].try_into()?);
+        let mod_state = u16::from_le_bytes(response[86..88].try_into()?);
+        let number_channels = u16::from_le_bytes(response[88..90].try_into()?);
+        Ok((
+            serial_number,
+            model_number,
+            hardware_type,
+            format!("{}.{}.{}", firmware_major, firmware_interim, firmware_minor),
+            hardware_version,
+            mod_state,
+            number_channels,
+        ))
     }
 }
